Added extended classes to handle complex object like Customer with addresses or orders with customers containing addresses Added a way to save X number of random customers with X numbe of random offers with X number of random products Added tests for Customers, Orders, Addresses and Products.

+++ src/test/java/pojos/cuustomeraddresses/CustomerAddressTests.java
@@ -34,7 +34,7 @@ public class CustomerAddressTests {
 
         for (CustomerAddress customerAddress :
                 customerAddresses) {
-            Long id = customerAddress.getId();
+            Integer id = customerAddress.getId();
             Assert.assertNotNull("Address is null for customerAddress with id:" + id,
                     customerAddress.getAddress());
             Assert.assertNotNull("City is null for customerAddress with id:" + id,
